feat. 공통다이얼로그
검색 창 분리 및 주소 컴포넌트 분리
