--- src/component/bgtcd/hi.js
@@ -1,1 +1,0 @@
-afds
